fix(mcp): split withdrawToAddress into "use node" file
`convex/mcp/desks.ts` contained an `internalAction` that dynamic-imports
`@coinbase/cdp-sdk` (and now also viem for the PR-A simulate), but the
file is missing the `"use node"` directive. `npx convex dev` fails to
bundle the CDP SDK's `crypto` import for the default isolate runtime.

This was a latent bug from PR #155 (the original Phase 6 withdraw flow);
PR-A (#157) only widened its surface by adding viem simulation alongside
the existing CDP call. Both modules need node.

Fix follows the established pattern in `dealsEscrow.ts` and
`tradersEscrow.ts`: keep the query + mutation in `desks.ts` (default
runtime), move the action into a sibling `desksEscrow.ts` with
`"use node"` at the top. The three address helpers (`normalizeAddress`,
`isAllowlisted`, `startOfUtcDay`) are exported from `desks.ts` and
imported by `desksEscrow.ts` — pure functions cross-runtime safely.

http.ts route pointer updated to `internal.mcp.desksEscrow.withdrawToAddress`.
