+++ convex/_generated/api.d.ts
@@ -40,9 +40,12 @@ import type * as mcp_approvals from "../mcp/approvals.js";
 import type * as mcp_deals from "../mcp/deals.js";
 import type * as mcp_dealsEscrow from "../mcp/dealsEscrow.js";
 import type * as mcp_desks from "../mcp/desks.js";
+import type * as mcp_desksEscrow from "../mcp/desksEscrow.js";
 import type * as mcp_httpHelpers from "../mcp/httpHelpers.js";
+import type * as mcp_limits from "../mcp/limits.js";
 import type * as mcp_outcomes from "../mcp/outcomes.js";
 import type * as mcp_requests from "../mcp/requests.js";
+import type * as mcp_simulate from "../mcp/simulate.js";
 import type * as mcp_traders from "../mcp/traders.js";
 import type * as mcp_tradersEscrow from "../mcp/tradersEscrow.js";
 import type * as mcpApiKeys from "../mcpApiKeys.js";
@@ -108,9 +111,12 @@ declare const fullApi: ApiFromModules<{
   "mcp/deals": typeof mcp_deals;
   "mcp/dealsEscrow": typeof mcp_dealsEscrow;
   "mcp/desks": typeof mcp_desks;
+  "mcp/desksEscrow": typeof mcp_desksEscrow;
   "mcp/httpHelpers": typeof mcp_httpHelpers;
+  "mcp/limits": typeof mcp_limits;
   "mcp/outcomes": typeof mcp_outcomes;
   "mcp/requests": typeof mcp_requests;
+  "mcp/simulate": typeof mcp_simulate;
   "mcp/traders": typeof mcp_traders;
   "mcp/tradersEscrow": typeof mcp_tradersEscrow;
   mcpApiKeys: typeof mcpApiKeys;
